Fase 4 do projeto

+++ app/Http/Controllers/ProjectController.php
@@ -12,6 +12,7 @@ namespace CodeProject\Http\Controllers;
 use CodeProject\Repositories\ProjectRepository;
 use CodeProject\Services\ProjectService;
 use Illuminate\Http\Request;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -32,23 +33,60 @@ class ProjectController extends Controller
         $this->service = $service;
     }
 
-    public function  index(){
-        return $this->repository->with(['owner','client'])->all();
+    public function index()
+    {
+        return $this->repository->findWhere(['owner_id' => Authorizer::getResourceOwnerId()]);
     }
 
-    public function storage(Request $request){
+    public function store(Request $request)
+    {
         return $this->service->create($request->all());
     }
 
-    public function show($id){
+    public function show($id)
+    {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->show($id);
     }
 
-    public function destroy($id){
+    public function destroy($id)
+    {
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->destroy($id);
     }
-    public function update(Request $request, $id){
-        return $this->service->update($request->all(),$id);
+
+    public function update(Request $request, $id)
+    {
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
+        return $this->service->update($request->all(), $id);
+    }
+
+    public function checkProjectOwner($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    public function checkProjectMember($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    public function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)) {
+            return true;
+        }
+
+        return false;
+
     }
 
 }
