+++ app/Http/Controllers/ProjectMembersController.php
@@ -24,41 +24,63 @@ class ProjectMembersController extends Controller
      * @var ProjectNoteService
      */
     private $service;
+    /**
+     * @var ProjectController
+     */
+    private $projectController;
 
-    public function __construct(ProjectMembersRepository $repository, ProjectMembersService $service)
+    public function __construct(ProjectMembersRepository $repository, ProjectMembersService $service, ProjectController $projectController)
     {
         $this->repository = $repository;
-
         $this->service = $service;
+        $this->projectController = $projectController;
     }
 
     public function index($id)
     {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->index($id);
     }
 
-    public function storage(Request $request)
+    public function store($id, Request $request)
     {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->addMember($request->all());
     }
 
     public function show($id, $memberId)
     {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->repository->findWhere(['project_id' => $id, 'id' => $memberId]);
     }
 
     public function destroy($id, $memberId)
     {
+        if ($this->projectController->checkProjectOwner($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->removeMember($id, $memberId);
     }
 
     public function member($id, $userId)
     {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->isMember($id, $userId);
     }
 
     public function update(Request $request, $id, $memberId)
     {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->update($request->all(), $memberId);
     }
