+++ app/Http/Controllers/ProjectTaskController.php
@@ -24,32 +24,56 @@ class ProjectTaskController extends Controller
      * @var ProjectNoteService
      */
     private $service;
+    /**
+     * @var ProjectController
+     */
+    private $projectController;
 
-    public function __construct(ProjectTaskRepository $repository, ProjectTaskService $service)
+    public function __construct(ProjectTaskRepository $repository, ProjectTaskService $service, ProjectController $projectController)
     {
         $this->repository = $repository;
-
         $this->service = $service;
+        $this->projectController = $projectController;
     }
 
-    public function  index($id){
+    public function index($id)
+    {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->repository->findWhere(['project_id' => $id]);
     }
 
-    public function storage(Request $request){
+    public function store($id, Request $request)
+    {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->create($request->all());
     }
 
-    public function show($id, $taskId){
+    public function show($id, $taskId)
+    {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->repository->findWhere(['project_id' => $id, 'id' => $taskId]);
     }
 
-    public function destroy($id, $taskId){
+    public function destroy($id, $taskId)
+    {
+        if ($this->projectController->checkProjectOwner($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
         return $this->service->destroy($taskId);
     }
 
-    public function update(Request $request, $id, $taskId){
-        return $this->service->update($request->all(),$taskId);
+    public function update(Request $request, $id, $taskId)
+    {
+        if ($this->projectController->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
+        return $this->service->update($request->all(), $taskId);
     }
 
 }
